separate various interfaces
