+++ libs/interfaces/cartDrawer.ts
@@ -1,0 +1,7 @@
+export interface CartDrawerInterface {
+    placement: string | any
+    isOpen?: boolean
+    onOpen?: () => void
+    onClose?: () => void
+    onToggle?: () => void
+}
