+++ libs/interfaces/cartItem.ts
@@ -1,0 +1,5 @@
+export interface CartItemInterface {
+    id: string
+    name: string
+    quantity: number
+}
